quickstart-connector moudle add
