+++ quickstart-connector/src/main/java/com/streamxhub/streamx/flink/quickstart/connector/KafkaSimpleJavaApp.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2019 The StreamX Project
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.streamxhub.streamx.flink.quickstart.connector;
+
+import com.streamxhub.streamx.flink.connector.kafka.source.KafkaJavaSource;
+import com.streamxhub.streamx.flink.connector.kafka.source.KafkaRecord;
+import com.streamxhub.streamx.flink.core.StreamEnvConfig;
+import com.streamxhub.streamx.flink.core.scala.StreamingContext;
+import org.apache.flink.api.common.functions.MapFunction;
+import org.apache.flink.streaming.api.datastream.DataStream;
+
+/**
+ * @author benjobs
+ */
+public class KafkaSimpleJavaApp {
+
+    public static void main(String[] args) {
+
+        StreamEnvConfig envConfig = new StreamEnvConfig(args, null);
+
+        StreamingContext context = new StreamingContext(envConfig);
+
+        DataStream<String> source = new KafkaJavaSource<String>(context)
+                .getDataStream()
+                .map((MapFunction<KafkaRecord<String>, String>) KafkaRecord::value);
+
+        source.print();
+
+        context.start();
+    }
+
+}
