--- quickstart-connector/src/main/java/com/streamxhub/streamx/test/flink/java/bean/Behavior.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2019 The StreamX Project
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.streamxhub.streamx.test.flink.java.bean;
-
-import lombok.Data;
-
-/**
- * @author benjobs
- */
-@Data
-public class Behavior {
-    private String user_id;
-    private Long item_id;
-    private Long category_id;
-    private String behavior;
-    private Long ts;
-}
