+++ quickstart-connector/src/main/scala/com/streamxhub/streamx/flink/quickstart/connector/MyDataSource.scala
@@ -1,0 +1,57 @@
+/*
+ * Copyright (c) 2019 The StreamX Project
+ * <p>
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package com.streamxhub.streamx.flink.quickstart.connector
+
+import com.streamxhub.streamx.flink.quickstart.connector.bean.Entity
+import org.apache.flink.streaming.api.functions.source.SourceFunction
+
+import scala.util.Random
+
+
+class MyDataSource extends SourceFunction[Entity] {
+
+  private[this] var isRunning = true
+
+  override def cancel(): Unit = this.isRunning = false
+
+  val random = new Random()
+
+  var index = 0
+
+  override def run(ctx: SourceFunction.SourceContext[Entity]): Unit = {
+    while (isRunning && index <= 1000001) {
+      index += 1
+
+      val entity = new Entity()
+      entity.userId = System.currentTimeMillis()
+      entity.orderId = random.nextInt(100)
+      entity.orderStatus = random.nextInt(1)
+      entity.price = random.nextDouble()
+      entity.quantity = new Random().nextInt(10)
+      entity.cityId = 1
+      entity.siteId = random.nextInt(20)
+      entity.timestamp = System.currentTimeMillis()
+
+      ctx.collect(entity)
+    }
+  }
+
+}
