--- quickstart-connector/src/main/scala/com/streamxhub/streamx/test/flink/scala/datastream/KafkaSourceApp.scala
@@ -1,56 +1,0 @@
-/*
- * Copyright (c) 2019 The StreamX Project
- * <p>
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.streamxhub.streamx.test.flink.scala.datastream
-
-import com.streamxhub.streamx.flink.connector.kafka.source.KafkaSource
-import com.streamxhub.streamx.flink.core.scala.FlinkStreaming
-import org.apache.flink.api.java.utils.ParameterTool
-import org.apache.flink.api.scala._
-import org.apache.flink.streaming.api.scala.StreamExecutionEnvironment
-
-object KafkaSourceApp extends FlinkStreaming {
-
-  /**
-   * 用户可覆盖次方法...
-   *
-   */
-  override def ready(): Unit = super.ready()
-
-
-  override def config(env: StreamExecutionEnvironment,
-                      parameter: ParameterTool): Unit = {
-
-  }
-
-  override def handle(): Unit = {
-
-    //one topic
-    KafkaSource().getDataStream[String]()
-      .uid("kfkSource1")
-      .name("kfkSource1")
-      .map(x => {
-        x.value
-      })
-      .print()
-
-  }
-
-}
